requirements.txt
streamlit>=1.35.0
openai>=1.30.0
faiss-cpu>=1.8.0
pypdf>=4.2.0
python-docx>=1.1.2
numpy>=1.26.0
python-dotenv>=1.0.1
